feat: Convert hooks to HookHandler interface
